requirements.txt
pinecone==5.4.2
spacy==3.7.4
fastcoref==2.1.6
transformers==4.34.1
torch==2.7.0
ai71==0.0.23
sentence-transformers==4.1.0
bertopic==0.16.2
boto3==1.35.88
huggingface-hub==0.17.3
datasets==2.14.4
langchain==0.3.24
langchain-core==0.3.56
langchain-nvidia-ai-endpoints==0.3.9
langchain-text-splitters==0.3.8
python-dotenv==1.1.0